Changes to package structure
