+++ sayn/scaffolding/data/init_project/project.yaml
