+++ sayn/scaffolding/data/init_project/sample_settings.yaml
