+++ sayn/scaffolding/data/init_project/sql/dim_fighters.sql
