+++ sayn/scaffolding/data/init_project/sql/f_fighter_results.sql
