+++ sayn/scaffolding/data/init_project/sql/f_rankings.sql
