+++ sayn/tasks/dummy.py
@@ -1,16 +1,15 @@
 from .task import TaskRunner
-from ..utils.ui import UI
 
 
 class DummyTask(TaskRunner):
     def setup(self):
-        UI().debug("Setting up Dummy Task")
+        self.logger.debug("Setting up Dummy Task")
         return self.ready()
 
     def compile(self):
-        UI().debug("DummyTask compiling")
+        self.logger.debug("DummyTask compiling")
         return self.success()
 
     def run(self):
-        UI().debug("DummyTask running.")
+        self.logger.debug("DummyTask running.")
         return self.success()
