+++ sayn/tasks/python.py
@@ -1,25 +1,15 @@
 from .task import TaskRunner
-from ..utils.ui import UI
 
 
 class PythonTask(TaskRunner):
-    def __init__(self, name, task):
-        super(PythonTask, self).__init__(name, task)
-
-        # Add some convenient properties
-        self.default_db = self.sayn_config.default_db
-
-        # Add UI
-        self.logger = UI()
-
     def setup(self):
-        UI().debug("Setting up Python Task")
+        self.logger.debug("Setting up Python Task")
         return self.ready()
 
     def compile(self):
-        UI().debug("PythonTask compiling")
+        self.logger.debug("PythonTask compiling")
         return self.success()
 
     def run(self):
-        UI().debug("PythonTask running.")
+        self.logger.debug("PythonTask running.")
         return self.success()
