--- sayn/utils/singleton.py
@@ -1,18 +1,0 @@
-class SingletonMeta(type):
-    _instance = None
-
-    def __call__(self):
-        if self._instance is None:
-            self._instance = super().__call__()
-        return self._instance
-
-
-def singleton(class_):
-    instances = {}
-
-    def getinstance(*args, **kwargs):
-        if class_ not in instances:
-            instances[class_] = class_(*args, **kwargs)
-        return instances[class_]
-
-    return getinstance
