+++ sayn/utils/task_query.py
@@ -1,12 +1,6 @@
-from datetime import datetime
 import re
-from uuid import UUID, uuid4
 
-from ..utils.ui import UI
-from ..utils.dag import query as dag_query
-from ..utils.dag import topological_sort
-from ..tasks import Task
-from .config import read_project, read_dags, read_settings, get_tasks_dict
+from ..errors import TaskQueryError
 
 #####################################
 # Task query interpretation functions
@@ -21,10 +15,6 @@ RE_TASK_QUERY = re.compile(
         r")$"
     )
 )
-
-
-class TaskQueryError(Exception):
-    pass
 
 
 def _get_query_component(tasks, query):
@@ -101,81 +91,3 @@ def get_query(tasks, include=list(), exclude=list()):
         for operation, operands in (("include", include), ("exclude", exclude))
         for task, flags in operands.items()
     ]
-
-
-class SaynApp:
-    run_id: UUID = uuid4()
-    run_start_ts = datetime.now()
-    ui = None
-
-    tasks = dict()
-    dag = dict()
-
-    task_query = list()
-
-    connections = dict()
-    default_db = None
-
-    def __init__(self, debug):
-        self.ui = UI(run_id=self.run_id, debug=debug)
-        self.ui._set_config(stage_name="setup")
-
-    def setup(
-        self,
-        include=list(),
-        exclude=list(),
-        profile=None,
-        full_load=False,
-        start_dt=None,
-        end_dt=None,
-    ):
-        # Read the project configuration
-        self.set_project(read_project())
-        self.set_dags(read_dags(self.project.dags))
-        self.set_settings(read_settings())
-
-        # Set tasks and dag from it
-        tasks_dict = get_tasks_dict(self.project.presets, self.dags)
-        self.task_query = get_query(tasks_dict, include=include, exclude=exclude)
-        self.set_tasks(tasks_dict)
-
-        # Set settings and connections
-
-    def set_project(self, project):
-        self.project = project
-
-    def set_dags(self, dags):
-        self.dags = dags
-
-    def set_settings(self, settings):
-        self.settings = settings
-
-    def set_tasks(self, tasks):
-        self.dag = {
-            task["name"]: [p for p in task.get("parents", list())]
-            for task in tasks.values()
-        }
-
-        self._tasks_dict = {
-            task_name: tasks[task_name] for task_name in topological_sort(self.dag)
-        }
-
-        tasks_in_query = dag_query(self.dag, self.task_query)
-
-        for task_name, task in self._tasks_dict.items():
-            self.tasks[task_name] = Task(
-                task,
-                [self.tasks[p] for p in task.get("parents", list())],
-                task_name in tasks_in_query,
-                self.project.parameters,
-                list(),
-                self.ui,
-            )
-
-    def run(self):
-        for task_name, task in self.tasks.items():
-            task.run()
-
-    def compile(self):
-        for task_name, task in self.tasks.items():
-            task.compile()
